requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
frozenlist==1.3.0
google-api-core==2.8.2
google-auth==2.8.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
multidict==6.0.2
numpy==1.23.0
packaging==21.3
proto-plus==1.20.6
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.0
rsa==4.8
six==1.16.0
urllib3==1.26.9
yarl==1.7.2
